+++ src/tests.rs
@@ -11,6 +11,7 @@
 
 use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
+use std::collections::BTreeMap;
 use std::ffi::OsStr;
 use std::fmt::Display;
 use std::fs;
@@ -61,13 +62,13 @@ pub enum TestStatus {
     Warning { details: Option<String> },
 }
 
-/// Stores all of the available tests. Used solely for pretty printing purposes in `clap-validator
-/// list tests`.
+/// Stores all of the available tests and their descriptions. Used solely for pretty printing
+/// purposes in `clap-validator list tests`.
 #[derive(Debug, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct TestList {
-    pub plugin_library_tests: Vec<String>,
-    pub plugin_tests: Vec<String>,
+    pub plugin_library_tests: BTreeMap<String, String>,
+    pub plugin_tests: BTreeMap<String, String>,
 }
 
 /// An abstraction for a test case. This mostly exists because we need two separate kinds of tests
@@ -224,18 +225,13 @@ impl TestStatus {
 
 impl Default for TestList {
     fn default() -> Self {
-        let mut list = Self {
+        Self {
             plugin_library_tests: PluginLibraryTestCase::iter()
-                .map(|c| c.to_string())
+                .map(|c| (c.to_string(), c.description()))
                 .collect(),
-            plugin_tests: PluginTestCase::iter().map(|c| c.to_string()).collect(),
-        };
-
-        // These should already be sorted, but since the validator results are also sorted we'll add
-        // a sort here just in caseq
-        list.plugin_library_tests.sort_unstable();
-        list.plugin_tests.sort_unstable();
-
-        list
+            plugin_tests: PluginTestCase::iter()
+                .map(|c| (c.to_string(), c.description()))
+                .collect(),
+        }
     }
 }