+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- `clap-validator list tests [--json]` now includes test descriptions.
 - When a plugin supports text-to-value and/or value-to-text conversions for some
   but not all of its parameters, clap-validator will now include the names of
   the parameters and the failing inputs in the error message to help pinpoint